+++ packages/xstate-graph/test/graph.test.ts
@@ -13,9 +13,9 @@ import {
   getShortestPaths,
   toDirectedGraph,
   StatePath,
-  joinPaths,
   getSimplePaths
 } from '../src';
+import { joinPaths } from '../src/graph';
 
 function getPathsSnapshot(
   paths: Array<StatePath<any, EventObject>>
@@ -94,7 +94,8 @@ describe('@xstate/graph', () => {
   }
   type CondMachineEvents = { type: 'EVENT'; id: string } | { type: 'STATE' };
 
-  const condMachine = createMachine<CondMachineCtx, CondMachineEvents>({
+  const condMachine = createMachine({
+    types: {} as { context: CondMachineCtx; events: CondMachineEvents },
     initial: 'pending',
     context: {
       id: undefined
@@ -199,7 +200,7 @@ describe('@xstate/graph', () => {
       );
     });
 
-    it('the initial state should have a zero-length path', () => {
+    it('the initial state should have a single-length path', () => {
       const shortestPaths = getShortestPaths(lightMachine);
 
       expect(
@@ -210,7 +211,7 @@ describe('@xstate/graph', () => {
             ).value
           )
         )!.steps
-      ).toHaveLength(0);
+      ).toHaveLength(1);
     });
 
     xit('should not throw when a condition is present', () => {
@@ -218,7 +219,8 @@ describe('@xstate/graph', () => {
     });
 
     it.skip('should represent conditional paths based on context', () => {
-      const machine = createMachine<CondMachineCtx, CondMachineEvents>({
+      const machine = createMachine({
+        types: {} as { context: CondMachineCtx; events: CondMachineEvents },
         initial: 'pending',
         context: {
           id: 'foo'
@@ -361,7 +363,7 @@ describe('@xstate/graph', () => {
       );
     });
 
-    it('should return a single empty path for the initial state', () => {
+    it('should return a single-length path for the initial state', () => {
       expect(
         getSimplePaths(lightMachine).find((p) =>
           p.state.matches(
@@ -379,7 +381,7 @@ describe('@xstate/graph', () => {
             ).value
           )
         )!.steps
-      ).toHaveLength(0);
+      ).toHaveLength(1);
       expect(
         getSimplePaths(equivMachine).find((p) =>
           p.state.matches(
@@ -397,7 +399,7 @@ describe('@xstate/graph', () => {
             ).value
           )
         )!.steps
-      ).toHaveLength(0);
+      ).toHaveLength(1);
     });
 
     it('should return value-based paths', () => {
@@ -408,7 +410,8 @@ describe('@xstate/graph', () => {
         type: 'INC';
         value: number;
       }
-      const countMachine = createMachine<Ctx, Events>({
+      const countMachine = createMachine({
+        types: {} as { context: Ctx; events: Events },
         id: 'count',
         initial: 'start',
         context: {
@@ -568,7 +571,8 @@ it('shortest paths for transition functions', () => {
 
 describe('filtering', () => {
   it('should not traverse past filtered states', () => {
-    const machine = createMachine<{ count: number }>({
+    const machine = createMachine({
+      types: {} as { context: { count: number } },
       initial: 'counting',
       context: { count: 0 },
       states: {
@@ -635,14 +639,14 @@ it('should provide previous state for serializeState()', () => {
     }
   });
 
-  // Should be [0, 3]:
-  // 0 (a)
-  // 3 (a -> b -> c -> a)
+  // Should be [1, 4]:
+  // 1 (a)
+  // 4 (a -> b -> c -> a)
   expect(
     shortestPaths
       .filter((path) => path.state.matches('a'))
       .map((path) => path.steps.length)
-  ).toEqual([0, 3]);
+  ).toEqual([1, 4]);
 });
 
 it.each([getShortestPaths, getSimplePaths])(
@@ -667,10 +671,10 @@ it.each([getShortestPaths, getSimplePaths])(
       fromState: machine.resolveState(machine.resolveStateValue('b'))
     });
 
-    // Instead of taking 1 step to reach state 'b', there should
-    // exist a path that takes 0 steps
+    // Instead of taking 2 steps to reach state 'b' (A, B),
+    // there should exist a path that takes 1 step
     expect(
-      paths.find((path) => path.state.matches('b') && path.steps.length === 0)
+      paths.find((path) => path.state.matches('b') && path.steps.length === 1)
     ).toBeTruthy();
 
     // Instead of starting at state 'a', it should take > 0 steps to reach 'a'
@@ -710,6 +714,7 @@ describe('joinPaths()', () => {
     expect(pathToBAndC.steps.map((step) => step.event.type))
       .toMatchInlineSnapshot(`
       [
+        "xstate.init",
         "NEXT",
         "TO_C",
       ]